combine into one plugin with auth
